CDLS-03: restore defaults for service
