+++ src/main/java/com/codeless/api/automation/controller/ExceptionTranslator.java
@@ -1,6 +1,8 @@
 package com.codeless.api.automation.controller;
 
 import com.codeless.api.automation.exception.ApiException;
+import javax.validation.ValidationException;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ControllerAdvice;
 import org.springframework.web.bind.annotation.ExceptionHandler;
@@ -12,5 +14,10 @@ public class ExceptionTranslator extends ResponseEntityExceptionHandler {
   @ExceptionHandler(ApiException.class)
   public ResponseEntity<String> handleApiException(ApiException apiException) {
     return ResponseEntity.status(apiException.getStatusCode()).body(apiException.getMessage());
+  }
+
+  @ExceptionHandler(ValidationException.class)
+  public ResponseEntity<String> handleValidationException(ValidationException validationException) {
+    return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(validationException.getMessage());
   }
 }
