--- src/main/java/com/codeless/api/automation/security/domain/UserRole.java
@@ -1,14 +1,0 @@
-package com.codeless.api.automation.security.domain;
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-@Getter
-@RequiredArgsConstructor
-public enum UserRole {
-
-    ROLE_USER("ROLE_USER"),
-    ROLE_ADMIN("ROLE_ADMIN");
-
-    private final String value;
-}
