--- src/main/java/com/codeless/api/automation/security/domain/View.java
@@ -1,6 +1,0 @@
-package com.codeless.api.automation.security.domain;
-
-public class View {
-    public static class ALLOWED { }
-    public static class NOT_ALLOWED { }
-}
