--- src/main/java/com/codeless/api/automation/security/repository/UserRepository.java
@@ -1,10 +1,0 @@
-package com.codeless.api.automation.security.repository;
-
-import com.codeless.api.automation.security.domain.User;
-import org.springframework.data.repository.CrudRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface UserRepository extends CrudRepository<User, Long> {
-    User findByUsername(String name);
-}
