--- src/main/java/com/codeless/api/automation/security/service/UserService.java
@@ -1,7 +1,0 @@
-package com.codeless.api.automation.security.service;
-
-import com.codeless.api.automation.security.domain.User;
-
-public interface UserService {
-    User saveUser(User user);
-}
