+++ src/main/java/com/codeless/api/automation/service/security/UserService.java
@@ -1,0 +1,8 @@
+package com.codeless.api.automation.service.security;
+
+import com.codeless.api.automation.entity.security.UserDto;
+
+public interface UserService {
+
+  UserDto saveUser(UserDto userDto);
+}
